requirements.txt
pytest~=7.1.2
Flask==2.1.2
boto3==1.22.3
Pillow~=9.1.0
botocore==1.25.3
moviepy~=1.0.3